Add erubi/capture file containing Erubi::CaptureEngine, allowing <%|= and <%|== for capture (and escaping) blocks in templates
The <%|= and <%|== tags support constructions like:

  <%|= form do %>
    <input>
  <% end %>

This construction is not supported by ERB or Erubis syntax.  Rails
attempts to support it via <%=, but it does so via an ugly hack of
attempting to parse ruby source code with a regexp.  By using a
separate tag for this, Erubi can support the same construction
without the parsing problems.

An advantage of this implementation over Rails is it has built in
support for capturing, you don't need to have all of the methods
that are used in the templates do capturing separately.  However, a
disadvantage is using non-local jumps (being/rescue, throw/catch)
inside a template when capturing is used is not supported.

Add :bufval and :regexp options to support this.  Also, switch the
if/else chain to a case statement, and add an else clause that
calls handle, which raises an exception by default.

CaptureEngine overrides handle to handle |= and |== indicators.

Add :src option to specify initial template source code.