Support CaptureEngine in the tilt support
This allows the :capture option in the tilt support to use
CaptureEngine.  It also adds support for the :engine_class option,
for similarity to tilt's erubis support, allowing custom engines
to be used.

This also has the tilt support be based off Tilt::Template instead
of Tilt::ERBTemplate, which means it no longer requires ERB. It
turns out this simplifies the code quite a bit as well.