Fix escaping in erubi/capture_end, the setting was previously inverted (Fixes #10)
This wasn't caught previously because the tests are parameterized
and it wasn't obvious what the settings was.  The code was based
on the default erubi code which uses xor, but a positive xor leads
to unescaped output in the default code, but previously lead to
escaped output in capture_end.

While here, add a couple non-parameterized capture_end tests to make
the expected behavior obvious, as well as a parameterized test for
using a loop inside a method that gets injected into a <%|= and
<%|== block.