Fix regression where only first backslash/apostrophe was escaped
d152b2dec928e1050d8c722a207b172bf118f4fb broke pretty much all usage
of Erubi on websites because it only escapes the first backslash
or apostrophe in the text when escaping.  This removes the use
of String#[]= to handle escaping, moving to String#gsub! (or
gsub in the frozen case).