Remove unnecessary branch
text should always be a string here.  It is created via
input[pos, len], and pos should never be after the end of the
string.  The code above in the unless ch == '=' block assumes it
is a string, and the one place it is reassigned assigns it to the
empty string.