Avoid unnecessary defined? usage on Ruby 3+ when using the :ensure option
defined? was used previously because before Ruby 3, accessing a
undefined instance variable resulted in a verbose mode warning.
As of Ruby 3, it no longer does, so there is no reason to use
defined?.  This results in fewer VM instructions, so this will
improve performance, though the difference is unlikely to be
measurable.