Add chain_appends option to simplify VM instructions (Fixes #32)
On CRuby, `b = ""; b << "" << ""`  is faster than
`b = ""; b << ""; b << ""`, even though the result is
the same, because CRuby's stack-based virtual machine
adds an unnecessary pop/getlocal_* instruction combination
(getinstancevariable if using an instance variable instead
of a local variable).

Avoiding the unnecessary instructions can result in up
to a 15% improvement for a typical template.

Unfortunately, this optimization is not compatible with
code that swaps the buffer variable, and there is production
code that does that.  So this optimization is opt-in via
the chain_appends option.

The implementation approach keeps generated code without
chain_appends the same as before.  This makes it less likely
that this change will introduce regressions.

A with_buffer private method has been added to handle bufvar
managment based on the chain_appends setting. A
terminate_expression private method has also been added, for
ensuring any current expression in the generated code has
been terminated.  This is needed in the chain_appends case,
as expressions may not be terminated by default when using
chain_appends.

Co-authored-by: Jeremy Evans <code@jeremyevans.net>