Add `freeze_template_literals` option to avoid String#freeze
`"foo".freeze` generates an `opt_str_freeze` instruction while
`"foo"` when frozen string literals is enabled generated a
`putobject` instruction.

The performance difference between the two is not that big,
but it's still less work as it doesn't have to check wether
String#freeze was redefined.

It also generate much less native code with YJIT.