Use erb/escape for faster html escaping if available
ERB::Escape#html_escape will return input strings as is without
allocating new strings if no escaping is needed, which can result
in better performance for such strings. This shouldn't cause any
issues in Erubi, because the result of Erubi.h is only used to
concat to an existing string.