Add erubi/capture_block, supporting capturing block output via standard <%= and <%== tags
This implementation is based on @seki's recent presentations at RubyWorld
and RubyKaigi regarding an implementation for stdlib ERB. This implementation
has the following changes:

1. The Buffer class is a String subclass for better compatibility with code
   that calls string methods on the buffer.

2. This uses <<= instead of +=, so that we don't have to override Buffer#+,
   since for strings that should not mutate the receiver.

3. This supports escaped tags <%== by using the |= operator, and has Buffer#|
   do the escaping and concatenating (this is an extension to ERB added by
   Erubis and retained by Erubi).

4. This handles empty tags by treating them as containing the empty string
   for compatibility.

Unfortunately, the capture_block support isn't 100% compatible with the
standard support.  At least, behavior for multiple statements inside
<%= and <%== differs:

  <%= 1 ; 2 %>
  # default: 2
  # capture_block: 1

The capture_block support doesn't support the :escapefunc option. For
custom escaping functions, you'll need to use a Buffer subclass. It also
doesn't support the :chain_appends option (it ignores it if set).

This tries to make as many tests as possible pass with the capture_block
support.  Compared to capture_end, there are a few whitespace differences,
and :escape_capture and :yield_returns_block options are not supported.
Otherwise, this passes the same specs if you change <%|= and <%|== to
<%= and <%==.

I would guess the capture_block implementation is not as optimized as
the default implementation.  It has to do the same amount of work, plus
unnecessarily set the buffer variable to itself (technically, set it to
the result of #<< or #|) for every <%= or <%==. However, it seems like
a worthwhile tradeoff if you want to support block capturing.